backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
mariadb==1.1.10
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
sentence-transformers==3.1.1
